backend/requirements.txt
argon2-cffi==23.1.0
bleach==6.2.0
cffi==1.17.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.10.0
Flask-SQLAlchemy==3.1.1
Flask-User==1.0.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
passlib==1.7.4
pillow==11.0.0
PyJWT==2.3.0
pyotp==2.9.0
qrcode==8.0
requests==2.25.1
rsa==4.9
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.5
Werkzeug==3.1.3
WTForms==3.2.1
pycryptodome
